Visual representation of cinema, select and reserve seats
