+++ client/src/styles.css
@@ -5,4 +5,3 @@
 @import "CSS/movies.css";
 @import "CSS/tickets.css";
 @import "CSS/reservations.css";
-@import "CSS/seats.css"
